BREAKING: Use packed structs (#7)
#### Problem

Many of the structs in the SDK use `extern` layout. While this ensures
compatibility with C ABI, it can also be less performant for on-chain
programs.

#### Summary of changes

Move `PublicKey`, `Account.Data`, and `Rent` to be `packed` structs.
This is a breaking change because the underlying layout of many types
will change.